+++ src/shapes.ts
@@ -141,12 +141,21 @@ export class Text extends Shape {
     ]
   }
 
+  private computeFontSize(text: string, defaultFontSize: number, context: CanvasRenderingContext2D): number {
+    let defaultTextWidth = 0;
+    if (defaultFontSize < 1) {
+      context.font = Text.buildFont(this.style, 1, this.font);
+      defaultTextWidth = context.measureText(text).width * defaultFontSize;
+    } else defaultTextWidth = context.measureText(text).width;
+    if (defaultTextWidth >= this.boundingBox.size.x) return defaultFontSize * this.boundingBox.size.x / defaultTextWidth;
+    return defaultFontSize
+  }
+
   private automaticFontSize(context: CanvasRenderingContext2D): number {
     let fontsize = Math.min(this.boundingBox.size.y ?? Number.POSITIVE_INFINITY, this.fontsize ?? Number.POSITIVE_INFINITY);
     if (fontsize == Number.POSITIVE_INFINITY) fontsize = DEFAULT_FONTSIZE;
     context.font = Text.buildFont(this.style, fontsize, this.font);
-    if (context.measureText(this.text).width >= this.boundingBox.size.x) fontsize = fontsize * this.boundingBox.size.x / context.measureText(this.text).width;
-    return fontsize
+    return this.computeFontSize(this.text, fontsize, context)
   }
 
   private setRectOffsetX(): number {
@@ -203,7 +212,13 @@ export class Text extends Shape {
       context.resetTransform();
       context.translate(origin.x, origin.y);
       context.rotate(Math.PI / 180 * this.orientation);
-      if (this.isScaled) context.scale(Math.abs(contextMatrix.a), Math.abs(contextMatrix.d));
+      if (this.isScaled) {
+        if (this.fontsize < 1) {
+          context.scale(Math.abs(contextMatrix.a * this.fontsize), Math.abs(contextMatrix.d * this.fontsize));
+          context.font = Text.buildFont(this.style, 1, this.font);
+        }
+        else context.scale(Math.abs(contextMatrix.a), Math.abs(contextMatrix.d));
+      }
       this.write(writtenText, context);
       context.restore();
     }