+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.22.5]
+### Fix
+- Fix bug on small scaled text position
+
 ## [0.22.4]
 ### Fix
 - Fix bug on tooltip origin when mouse leaving while hovering a shape